requirements.txt
langchain==0.0.350
streamlit==1.29.0
openai==1.5.0
python-dotenv==1.0.0